app/api/authViews.py
#!/usr/bin/env python
#-*- coding:utf-8 -*-
'''''
Created on
@author: zhuangyan
'''
import datetime
import json
import os
import uuid

from bson import ObjectId
from flask import Flask, Response, jsonify, request, session
from flask_uploads import UploadSet, extension
from PIL import Image


from com.utils import encrypt_password
from app import mongo

from . import api

