requirements.txt
Flask>=0.12
flask_pymongo
flask_script
flask-uploads >= 0.2.1
Jinja2>=2.9
Werkzeug>=0.10
pillow==5.0
flask-jwt
